+++ frontend/src/components/settings/settings.jsx
@@ -1,20 +1,25 @@
-import React, { useState } from 'react'
-import friendsApi from '../../utils/API/fetchUser'
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import friendsApi from '../../utils/API/fetchUser';
+import fetchAuth from '../../utils/API/fetchAuth';
 
 import './settings.css'
+import Modal from '../modal/modal';
 
 const userApi = new friendsApi()
+const auth = new fetchAuth()
 export default function Settings() {
 
+    const navigate = useNavigate()
     const [contact, setContact] = useState({
         tel: ''
     })
     const [openInput, setOpenInput] = useState(true)
     const [message, setMessage] = useState({
         msg: '',
-        showMessage:false
+        showMessage: false
     })
-
+    const [openModal, setOpenModal] = useState(false)
     const handleChange = (e) => {
         setContact({
             tel: e.target.value
@@ -25,36 +30,53 @@ export default function Settings() {
         e.preventDefault()
         try {
             await userApi.addFriend(contact)
-        } catch ({response}) {
-            setMessage({msg: response.data.msg, showMessage: true})
-            console.log(response.data.msg);
+        } catch ({ response }) {
+            setMessage({ msg: response.data.msg, showMessage: true })
         }
         setOpenInput(false)
     }
 
+    const handleClose = () => {
+        setOpenModal(false)
+    }
+
+    const logOut = async () => {
+        try {
+            await auth.logOut()
+            localStorage.setItem('chat-session', JSON.stringify(false))
+            navigate('/login')
+        } catch (error) {
+            console.log(error)
+        }
+    }
+
     return (
-        <>
+        <Modal open={openModal} onClose={handleClose}>
+        {/* <> */}
             {
-                openInput 
+                openInput
                     ?
                     <nav className='setting-nav'>
                         <ul>
                             <li><button>New Group</button></li>
-                            <li><button onClick={()=>{setOpenInput(false)}}>Add user</button></li>
+                            <li><button onClick={() => { setOpenInput(false) }}>Add user</button></li>
                             <li><button>Setting</button></li>
+                            <li><button onClick={logOut} >logOut</button></li>
                         </ul>
                     </nav>
                     :
                     <form onSubmit={handleSubmit} className='form-request'>
+                        <h3>Contacts</h3>
                         <input
                             type="tel"
                             placeholder='tel'
                             onChange={handleChange}
                         />
-                        { message.showMessage ? message.msg : '' }
+                        <p>{message.showMessage ? message.msg : ''}</p>
                         <button type='submit'>Add</button>
                     </form>
             }
-        </>
+        {/* </> */}
+        </Modal>
     )
 }
